fetch Data from fetch Calls
You can not just call a fetch function inside a setState call and expected the result of fetch to be returned to setState and stored in it eventually.
Instead, you use the setState function inside the fetchHandler function and store the data in it yourself. i.e

WRONG: setState( fetchHandler(apiKey) )
CORRECT:
fetchHandler() {
...  data
    setState(data )
}

Ofcourse if you use the setState in a fetchHandler, then you have to pay attention if the fetchHandler is inside the useCallback or not. If it is wrapped in a callback then pass the setState as a dependancy of callback