Default Parameter issue solved
I was calling the fetchHandler directly in the onClick prop.
by default react passes the click object as the parameter in the onClick prop function.
Therefore instead of passing a null as the parameter to fetchHandler which should have been converted to a default value of washington,
I was actually passing the click function to my fetchHandler.
This meant neither did i have any access to my default parameter nor does the object classify as any string for the api call.