Checkpoint
The api call is made on every input submission but the weather data changes only if
- The response is valid.
However the invalid input is stored in the recent entry list bcz the form is independent of the current weather.

This makes me realize that maybe the form should be placed within the current weather component. Earlier i tried to put it in the index page, but that was not appropriate as the page components should be clean and lean. However, component nesting is not a problem