One solved one created
Solved the issue of invalid entries being updated as history.
However, there is a new issue that entries result in creation of two duplicate items immediately.
In the console, `use Effect running  is shown, which indicates that the 2nd render is due to change in use effect dependency.
in process of trouble shooting....