Problem highlighted - Kind of
the useEffect ran the fetchHandler 2nd time bcz of the change in its last location dependancy. This dependancy was actually the location name from the last weather search (the valid location returned by api).
Earlier this behaviour was needed to call the api as soon as the history list changes, however now the history list changes on the apiCall itself.
Earlier change in history resulted in api call, but not its the opposite. The api call updates the history.
Therefore there is no need to add the last location as the dependency of the useEffect to detect the changes in history list. as it changes on valid api responses only and the list is always up to datee
There was only one way the lastLocation changed, and that was if someone ran the fetch handler which would