Restructuring & app working as it was before (issues exist)
Changes:

Issue:  Multiple locations with the same ID can be added to the list
Fix: Filtering logic to be added in the dispatch function that checks if the search term is already present in the state. If its available then there just needs to be a rearrangement in the location array that too if the current search is not the last search term. This also means that the response from the search should not change the data in the infoPanel as it is the same and no rerendering is needed.