troubleshooting state access - 01
Couldnot access the recentLocations from inside the fetchHandler.
Then checked if other states like locationInput or the context State as a whole is accessible.
Found that no state is accessible from inside the fetchHandler, which explained why did it not realize the repetition of location history inputs.
I am guessing that the issue is caused bcz i did not pass the state as a dependency of the useCallback around fetchHandler