i was right but it doesn't help
Bcz of not declaring the state as a dependency, i was not able to access it.
But if i declare it then every time the state changes, the fetchHandler is recreated and the useEffect is called again.
This is a looping issue bcz i am changing the state from within my fetchHandler and at the same time adding the state in the dependency. :(
FIX:
I should take the code of state manipulation out of fetchHandler and only limit its body to data fetching.