changed my approach
not manipulating state from within the fetchHandler means getting the response out of fetchHanlder. This means returning something from fetchHandler like
fetchResponse=fetchHanlder()
and this caused issues.
On a second thought, i realized that i have done state manipulation in the fetch handlers in the past and it didn't create any issues.
The problem in my present case was that i was accessing the state from my fetch handler whereas in the past i only accessed the state-setting functions like dispatch, from inside my fetch handlers.
Therefore i have to change my code in such a way that only the fetch-handler is used to set-state and not to consume-state.