Got rid of bug
the check (!!matchingIndex) was changed to (matching index >0).
However, i feel like scan Location history should return a recentLocationArray even when it does not match.
So that the conditional filter code may not have to run in the reducer.
Also, this would mean that the reordered element's Id should be updated too.