Resolved another bug
changed the check from
(matchingIndex > 0)
to
(matchingIndex >= 0).

Result:
Earlier it didn't check for the matches if the matched location was at index 0. This meant that 'washington' was never detected to be present and more washingtons would be added in the array. Even if there is a washington present at index 1 besides 0, bcz the first detection of index would be '0'.