stuck
Sometimes an error shows up that the values that your trying to read (to map) are undefined. However, this randomly becomes ok to use.
I think the problem is with multiple render cycles which causes the state to not be available sometimes. Due to this, the empty state returns the error when i try to transform the api data from it