data passing - phase 01 complete
Issue Resolved: The issue of undefined segregated data was caused due to the empty state values during the first cycle of the app (before the dispatch function is called). The component could not receive any data in that cycle and returned errors. Therefore now the data is checked to be valid and a dummy data is being passed in case empty data is there.

Moreover, all the data is now being successfully checked and passed in all components, but the problem caused now is that no formatting functions like capitalizations or rounding off is applied on the values which disturb the layout. This needs to be taken care of