routing with navigator object - problem solved
Problem:
The useEffect ran after the component render and by that time the Link had the old value of location saved in the href. This caused the undefined page to load even when the useEffect changed the location object after the component render.
Solution:
the location object was made a state object. This caused the href value to update when the useEffect ran after the component render to change the location state. On state change the new location state was now transferred to the Link component and clicking it now would redirect us to the desired page with updated location.
Note:
1. navigator api doesn't work outside the useEffect hook.
2. setState needed to be added as a dependency of the useEffect hook