searchList component created but not working
clicks on the list items are not triggering the click handler.
This is because the list is shown only if the form input (sibling of list) is in-focus.
But to click the list item, user has to click in the list, thus outside of the form input.
This triggers the on-blur event which is called before the list click event therefore by the time we read the click event there is not list in the dom and the click event is therefore triggered on the element behind the list i.e week forecast card.
To solve the issue, i must transfer the list in the form input component.